+++ src/core/util/AnnexMethods.java
@@ -1,0 +1,20 @@
+package core.util;
+
+/**
+ * This class stores many diverse methods such as math ones.
+ * 
+ * @author Raph
+ *
+ */
+public final class AnnexMethods {
+	
+	public static float clamp(float value, float a, float b) {
+		/* Return value clamped between a and b. */
+		return Math.max(a, Math.min(b, value));	
+	} 
+	
+	
+	
+	
+
+}
