--- src/core/util/Bounds.java
@@ -1,13 +1,0 @@
-package core.util;
-
-/**
- * Class which represents bounds... @@@
- * 
- * @author Raph
- *
- */
-public class Bounds {
-
-	//@TODO
-	
-}
