+++ src/core/util/CollisionBox.java
@@ -1,0 +1,11 @@
+package core.util;
+
+/**
+ * 
+ * 
+ * @author Raph
+ *
+ */
+public class CollisionBox {
+
+}
