+++ src/core/util/Vector2.java
@@ -2,12 +2,14 @@ package core.util;
 
 /**
  * Simple structure which represent a 2-dimensional vector.
- * Contains convenient static variables which can be used.
+ * Contains:
+ * 		classic functions on vectors
+ * 		convenient static variables which can be used
  * 
  * @author Raph
  * 
  */
-public final class Vector2 {		// @@@ final?
+public final class Vector2 {
 
 	public float x;
 	public float y;
@@ -18,6 +20,21 @@ public final class Vector2 {		// @@@ final?
 	}
 		
 	
+	public Vector2 add(Vector2 vectorToAdd) {
+		/* Add two vectors. 
+		 * NOT IN PLACE
+		 */
+		return new Vector2(this.x + vectorToAdd.x, this.y + vectorToAdd.y);
+	}
+	
+	public Vector2 multiply(float factor) {
+		/* Return the product of this vector by the scalar factor. 
+		 * NOT IN PLACE
+		 */
+		return new Vector2(factor * x, factor * y);
+	}
+	
+	
 	
 	public static Vector2 zero	= new Vector2(0, 0);
 	public static Vector2 up	= new Vector2(0, 1);
@@ -25,4 +42,12 @@ public final class Vector2 {		// @@@ final?
 	public static Vector2 left	= new Vector2(-1, 0);
 	public static Vector2 right	= new Vector2(1, 0);
 	
+	
+	
+	
+	
+	@Override
+	public String toString() {
+		return "(" + x + ", " + y + ")";
+	}
 }
