--- src/core/util/specific/About this package
@@ -1,3 +1,0 @@
-
-This package is for very specific class used in one particular context.
-Please document each class and indicate where it is used.
