+++ src/test/ABOUT THIS PACKAGE
@@ -1,0 +1,2 @@
+
+This package is isolated from the rest and is used to proceed to standalone tests.
