+++ src/test/Draft.java
@@ -1,0 +1,35 @@
+package test;
+
+import core.util.Vector2;
+
+/**
+ * This class is simply used as a draft to try out various things.
+ * 
+ * @author Raph
+ *
+ */
+public class Draft {
+	public static void main(String[] args) {
+		
+		Vector2 mep = new Vector2(1, 2);
+		System.out.println(mep);
+		
+		Vector2 mep2 = Vector2.multiply(12, mep);
+		System.out.println(mep2);
+		
+		
+		
+		
+		
+		
+	}
+	
+	
+	
+	
+	
+	
+	
+
+}
+
